+++ aim/deploy/docker/configs.py
@@ -1,0 +1,4 @@
+OPS_DIR = '_ops'
+SERVER_FILES_DIR = '_server_files'
+DOCKER_FILE_TEMPLATE = 'Dockerfile.py'
+DOCKERFILE_BUILD = 'Dockerfile'